enable setting atomic density path as code extra
The chargemol input files contain a path to a directory with atomic
densities. Unfortunately this makes chargemol input files completely
unportable, and means that one needs to change the input file if one
changes computer.

This PR adds a possibility to set a "DDEC_ATOMIC_DENSITIES_DIRECTORY"
extra on the ddec code.
The CalcJob input plugin will look for this extra if the location of the
densities is not otherwise specified.

Furthermore, we now validate the input dictionary using voluptuous
schemas.