validate input and set densities dir
We now validate the input dictionary using voluptuous schemas.

The chargemol input files contain a path to a directory with atomic
densities. Unfortunately this makes chargemol input files completely
unportable, and means that one needs to change the input file if one
changes computer.

This commit makes it possible to omit the explicit path to the atomic
densities file by setting a 'DDEC_ATOMIC_DENSITIES_DIRECTORY' extra on
the ddec Code instance.

Co-authored-by: Aliaksandr Yakutovich <yakutovicha@gmail.com>