symlink atomic_densities directory
Before, the absolute path of the atomic densities directory was written
into the input file, which was problematic for the hash-based comparison
of inputs in aiida-testing.

This commit creates a symlink to the atomic_densities directory in the
run folder, thus making sure that the *content* of the directory is what
matters, rather than its path.